Crowd-Funding-Web-app/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
django-allauth
requests
cryptography
Pillow